Added necessary code to generate an SVG file of the project dependencies.
Having a visual version of that is much better.
Plus, I'm likely to grow that number so I'm thinking it's going to help me when I need to regenerate the projects on Launchpad since the order is important for them (but they do not manage them at all.)
Maybe I'll even create a graphical tool which allows us to send the source packages with a simple click and in the right order with us just saying what changed.