Added a coverage script that I can use with all the projects.
Updated the mk script to handle coverage as well (-c/--coverage).
Added a "--clean" option to the mk script.
Updated the failure message to mention Tests & Coverage as well.
Renamed the SANITIZE_OPTIONS as just SANITIZE.
Print a message when the SANITIZE:BOOL=ON is used.
Removed the `${project_name}_` from the `COVERAGE` option.