Changed name of external test variable.
The test variable is now used along the -t command line option, so we
named it PROJECT_TEST_ARGS instead of PROJECT_COVERAGE_ARGS. It's used
to run the tests, not the coverage (although it is still used in the
coverage when running the tests in that script).

Also prevent --coverage when there is no tests directory.

Cleaned up the command line options (help, order, etc).