Moved the tmp & coverage directories.
I used a tmp directory in the source directory. That's not practical when doing a search, though (since you can end up finding results in those many unwanted files).

The new location is the same as the ${CMAKE_BINARY_DIR}. There we create a coverage and a tmp folder. The coverage folder is used when running the test for coverage information. The tmp folder is passed down to the tests so they can create files there when necessary. We now have a --tmp-dir option in the snapcatch2 so all our tests will support a temporary directory as expected.