Use the PROCESSORS defined by the user.
By default we use `nproc` which is fine 99% of the time, except in the huge
snapwebsites folder where it can be too many processors (if you do not have
enough RAM to make it all work as expected).

The user may create a ~/.build-snap.rc and define the PROCESSORS parameter in
that file to limit the number of processors the mk script should use. Now the
mk script loads that parameter.