Added a dependency on debian/changelog.
Whenever we build a project anew, we gather the version from the
debian/changelog file. The problem has been that updating the changelog
did not update the version in the project. We just have to nudge the
system to get that to happen, but you have to remember to reconfigure
(i.e. re-run cmake).

This change is expected to make that happen automatically.