Added the --tg flag to test in gdb.
By default the tests run freely. If there is a crash (SEGV, in most cases),
then you can capture those errors in gdb which in turn can tell you exactly
where it happened (i.e. perfect stack trace). This option starts the test
in gdb in a completely automated manner and with the same options as the
regular -t|--test option(s) would. It also supports a list of test names.