Made the -l work properly.
In the previous version, the -l would be applied to the compile and then
the test (and possibly some other things). Now, the script does two things:
1. it sends the output to a log file
2. it uses a sub-shell so a single less is required

Further, I changed the tabs in 4 spaces.
I check more of the command line options for conflicts.
Removed a variable that was not used.
A few other clean ups