Added locks in coverage & mk scripts.
I realized that if we have multiple consoles in the same directory, it's
easy to make the mistake of running the same ./mk or ./coverage in parallel
(which won't work).

Moved the cloc calls inside a for loop.
Removed the timeout of the cloc tool (although it may not serve us well).