Unlock must be defined after we obtained the lock.
If a script is still running, we fail obtaining the lock and if the EXIT is
used early on, the file gets deleted so the next attempt will go through when
the former script is still running. So we have a small "race condition"
window but we just can't avoid it. The trap is likely to always work, though.