Added support for sanitizer options.
At the moment, these are not cummulative although they should be. The issue
I ran into are memory leask from the OpenSSL library. To avoid errors running
the tests with the sanitizer, I have a list of the leaking functions. That
works, but once using that library in another higher level test, I run in
similar issues and would need these sanitizer options to be included so
the test passes.