Use a dict to avoid duplicates.
The cmake environment generates a deps.make file with the name of each
project and its dependencies found within the Snap! C++ endeavor. That
file does not properly get deleted when the cmake scripts run a second
time and re-generate the clean-dependencies.svg file.

The python script will now avoid the duplication so the SVG file looks
as expected. However, what we really need to fix is the generation of
the deps.make file, so I added a note in our TODO file.