Added macro to run our unit tests.
I'd like our build server to run our unit tests so when a project builds,
we then run the tests before we create the packages. If the tests fail,
then we have to go back to them to make them pass. This should force me
to better keep the code up to date.

Fixed the error detecting the BUILD and dist directories in the mk script.