Fixed the PROJECT_TEST_ARGS.
These are transmitted as a list which needs to be converted to a string to
work properly in a command line. So I added the necessary to convert the
list to a string.

I think that the --extra-options test (to avoid the shift) is not required
when used from the custom command (because we always pass at least a string
with a space) but since it is a separate script, I think it's safer to keep
that change too.