Prevent -l with -c.
The less and coverage functionality are not compatible in the current
version. The output is saved in a file, though, thus it's possible to
check the file with less ... after.

Also moved the launchpad confirmation after testing the parameters.

Fixed a parameter check for non-empty, the `test ...` keyword was missing.