Somehow, had to add --ignore-errors mismatch
There seems to be a bug and now I hit it with one of the projects. A
function ending point has one specific line position. What happens is that
in some cases, the coverage system detects two different position and picks
one of them randomly (newer versions will pick the largest, but in either
case it rather random).

The command line option prevents the whole thing from not working at all.
So far, though I have not noticed any issues in the resulting output. So
I think I'm good.