Replace lock file location.
The /run/user/${UID} location is only available if a user is currently
logged in. Although the ./mk script is expected to be used by developers
at times we may want to start it from a script (like on our system that
runs all the tests automatically). In that case, that folder does not
always exist since we run the script using CRON.

This uses a tmp directory inside the BUILD folder which is creates as
required. This way, it is also specific to one build making it possible
for a user to have multiple builds running in parallel as long as they
have multiple separate copies of the source code.